Agrego el cambio a la parte de escribir e intentar parsear el archivo para editar el mapa

+++ src_files/src/src_editor/ClickableLabel.cpp
@@ -1,13 +1,13 @@
 #include "ClickableLabel.h"
 
 ClickableLabel::ClickableLabel(std::vector<std::vector<char>>* grilla_terrenos,
-                               std::vector<std::vector<uint8_t>>* grilla_texturas,
+                               std::vector<std::vector<int>>* grilla_texturas,
                                std::vector<std::tuple<int, int>>& centros_ubicados,
                                int &cant_jugadores,
                                int &cant_filas_grillas,
                                int &cant_columnas_grillas,
                                char* terreno_seleccionado,
-                               uint8_t* textura_seleccionada,
+                               int* textura_seleccionada,
                                QString* path_textura,
                                QWidget* parent,
                                Qt::WindowFlags f) :
