+++ src_files/src/src_editor/ClickableLabel.h
@@ -20,24 +20,24 @@ class ClickableLabel : public QLabel {
 Q_OBJECT
 private:
     std::vector<std::vector<char>>* grilla_terrenos = nullptr;
-    std::vector<std::vector<uint8_t>>* grilla_texturas = nullptr;
+    std::vector<std::vector<int>>* grilla_texturas = nullptr;
     std::vector<std::tuple<int, int>>& centros_ubicados;
     int &cant_jugadores;
     int &cant_filas_grillas;
     int &cant_columnas_grillas;
     char* terreno_seleccionado = nullptr;
-    uint8_t* textura_seleccionada = nullptr;
+    int* textura_seleccionada = nullptr;
     QString* path_textura_seleccionada = nullptr;
 
 public:
     ClickableLabel(std::vector<std::vector<char>>* grilla_terrenos,
-                   std::vector<std::vector<uint8_t>>* grilla_texturas,
+                   std::vector<std::vector<int>>* grilla_texturas,
                    std::vector<std::tuple<int, int>>& centros_ubicados,
                    int &cant_jugadores,
                    int &cant_filas_grillas,
                    int &cant_columnas_grillas,
                    char* terreno_seleccionado,
-                   uint8_t* textura_seleccionada,
+                   int* textura_seleccionada,
                    QString* path_textura,
                    QWidget* parent = Q_NULLPTR,
                    Qt::WindowFlags f = Qt::WindowFlags());
